+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 - Scroll to top when loading another radar #19
+- Radar title in the lists should show full text with multiple lines #21
 
 ## [1.2.0(11)] - 2018-07-26
 ### Added